Add HTMX for smoother UX
Trying out HTMX. Instead of reloading the entire page when a download is
in progress, we only poll the `div` that contains a job that is being
worked on, with HTMX.

Also adds a nice throbbing animation on the icon as long as the polling
is happening.

Since the HTMX artifact is quite heavy with 45K uncompressed, add gzip
compression for all requests.